Suggest changing the sorting default to be "mean"
Consider a set of 10 samples in two groups A (3 samples) and B (7 samples). 
if a species makes up 100 % in all samples of group A, but is not present in any samples of B. 
The median for this species will be ZERO, and when we then sort by the median this species would not occur in the plot!
Sorting by the mean (as is currently implemented in the amp_heatmap) would "solve" this problem.