Update amp_rabund.R
Updated the input variable sort.by to accept lowercase inputs for "mean", "median" and "total" to be consistent with amp_heatmap.

Maybe the entire package needs a consistency check as I can see that deafault settings for scale.seq is 10000 in amp_rabund but 100 in amp_heatmap. And possibly an updated tutorial to match recommendations.